+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/impl/AuthAuthorityDaoSpringJdbc.java
@@ -8,32 +8,49 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import javax.sql.DataSource;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.util.Optional;
+import java.util.UUID;
 
 public class AuthAuthorityDaoSpringJdbc implements AuthAuthorityDao {
 
-  private final DataSource dataSource;
+    private final DataSource dataSource;
 
-  public AuthAuthorityDaoSpringJdbc(DataSource dataSource) {
-    this.dataSource = dataSource;
-  }
+    public AuthAuthorityDaoSpringJdbc(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
 
-  @Override
-  public void create(AuthorityEntity... authority) {
-    JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
-    jdbcTemplate.batchUpdate(
-        "INSERT INTO authority (user_id, authority) VALUES (? , ?)",
-        new BatchPreparedStatementSetter() {
-          @Override
-          public void setValues(PreparedStatement ps, int i) throws SQLException {
-            ps.setObject(1, authority[i].getUserId());
-            ps.setString(2, authority[i].getAuthority().name());
-          }
+    @Override
+    public void create(AuthorityEntity... authority) {
+        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
+        jdbcTemplate.batchUpdate(
+                "INSERT INTO authority (user_id, authority) VALUES (? , ?)",
+                new BatchPreparedStatementSetter() {
+                    @Override
+                    public void setValues(PreparedStatement ps, int i) throws SQLException {
+                        ps.setObject(1, authority[i].getUser().getId());
+                        ps.setString(2, authority[i].getAuthority().name());
+                    }
 
-          @Override
-          public int getBatchSize() {
-            return authority.length;
-          }
-        }
-    );
-  }
+                    @Override
+                    public int getBatchSize() {
+                        return authority.length;
+                    }
+                }
+        );
+    }
+
+    @Override
+    public Optional<AuthorityEntity> findById(UUID id) {
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<AuthorityEntity> findByUserId(UUID userId) {
+        return Optional.empty();
+    }
+
+    @Override
+    public void delete(AuthorityEntity... authority) {
+
+    }
 }
