+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/impl/AuthUserDaoSpringJdbc.java
@@ -15,45 +15,55 @@ import java.util.UUID;
 
 public class AuthUserDaoSpringJdbc implements AuthUserDao {
 
-  private final DataSource dataSource;
+    private final DataSource dataSource;
 
-  public AuthUserDaoSpringJdbc(DataSource dataSource) {
-    this.dataSource = dataSource;
-  }
+    public AuthUserDaoSpringJdbc(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
 
-  @Override
-  public AuthUserEntity create(AuthUserEntity user) {
-    JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
-    KeyHolder kh = new GeneratedKeyHolder();
-    jdbcTemplate.update(con -> {
-      PreparedStatement ps = con.prepareStatement(
-          "INSERT INTO \"user\" (username, password, enabled, account_non_expired, account_non_locked, credentials_non_expired) " +
-              "VALUES (?,?,?,?,?,?)",
-          Statement.RETURN_GENERATED_KEYS
-      );
-      ps.setString(1, user.getUsername());
-      ps.setString(2, user.getPassword());
-      ps.setBoolean(3, user.getEnabled());
-      ps.setBoolean(4, user.getAccountNonExpired());
-      ps.setBoolean(5, user.getAccountNonLocked());
-      ps.setBoolean(6, user.getCredentialsNonExpired());
-      return ps;
-    }, kh);
+    @Override
+    public AuthUserEntity create(AuthUserEntity user) {
+        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
+        KeyHolder kh = new GeneratedKeyHolder();
+        jdbcTemplate.update(con -> {
+            PreparedStatement ps = con.prepareStatement(
+                    "INSERT INTO \"user\" (username, password, enabled, account_non_expired, account_non_locked, credentials_non_expired) " +
+                            "VALUES (?,?,?,?,?,?)",
+                    Statement.RETURN_GENERATED_KEYS
+            );
+            ps.setString(1, user.getUsername());
+            ps.setString(2, user.getPassword());
+            ps.setBoolean(3, user.getEnabled());
+            ps.setBoolean(4, user.getAccountNonExpired());
+            ps.setBoolean(5, user.getAccountNonLocked());
+            ps.setBoolean(6, user.getCredentialsNonExpired());
+            return ps;
+        }, kh);
 
-    final UUID generatedKey = (UUID) kh.getKeys().get("id");
-    user.setId(generatedKey);
-    return user;
-  }
+        final UUID generatedKey = (UUID) kh.getKeys().get("id");
+        user.setId(generatedKey);
+        return user;
+    }
 
-  @Override
-  public Optional<AuthUserEntity> findById(UUID id) {
-    JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
-    return Optional.ofNullable(
-        jdbcTemplate.queryForObject(
-            "SELECT * FROM \"user\" WHERE id = ?",
-            AuthUserEntityRowMapper.instance,
-            id
-        )
-    );
-  }
+    @Override
+    public Optional<AuthUserEntity> findById(UUID id) {
+        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
+        return Optional.ofNullable(
+                jdbcTemplate.queryForObject(
+                        "SELECT * FROM \"user\" WHERE id = ?",
+                        AuthUserEntityRowMapper.instance,
+                        id
+                )
+        );
+    }
+
+    @Override
+    public Optional<AuthUserEntity> findByUsername(String username) {
+        return Optional.empty();
+    }
+
+    @Override
+    public void delete(AuthUserEntity user) {
+
+    }
 }
