+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/mapper/UdUserEntityRowMapper.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.data.mapper;
 
-import guru.qa.niffler.data.entity.userdata.UserEntity;
+import guru.qa.niffler.data.entity.userdata.UdUserEntity;
 import guru.qa.niffler.model.CurrencyValues;
 import org.springframework.jdbc.core.RowMapper;
 
@@ -8,7 +8,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.UUID;
 
-public class UdUserEntityRowMapper implements RowMapper<UserEntity> {
+public class UdUserEntityRowMapper implements RowMapper<UdUserEntity> {
 
   public static final UdUserEntityRowMapper instance = new UdUserEntityRowMapper();
 
@@ -16,8 +16,8 @@ public class UdUserEntityRowMapper implements RowMapper<UserEntity> {
   }
 
   @Override
-  public UserEntity mapRow(ResultSet rs, int rowNum) throws SQLException {
-    UserEntity result = new UserEntity();
+  public UdUserEntity mapRow(ResultSet rs, int rowNum) throws SQLException {
+    UdUserEntity result = new UdUserEntity();
     result.setId(rs.getObject("id", UUID.class));
     result.setUsername(rs.getString("username"));
     result.setCurrency(CurrencyValues.valueOf(rs.getString("currency")));
